ref: avoid circular dependencies
