+++ tools/ci_connector_ops/ci_connector_ops/pipelines/builds/__init__.py
@@ -3,11 +3,13 @@
 #
 """This module groups factory like functions to dispatch builds steps according to the connector language."""
 
+from __future__ import annotations
+
 import platform
-from typing import Optional, Tuple
+from typing import TYPE_CHECKING, Optional, Tuple
 
 import anyio
-from ci_connector_ops.pipelines.bases import ConnectorReport, StepResult
+from ci_connector_ops.pipelines.bases import ConnectorReport
 from ci_connector_ops.pipelines.builds import common, java_connectors, python_connectors
 from ci_connector_ops.pipelines.contexts import ConnectorContext
 from ci_connector_ops.utils import ConnectorLanguage
@@ -15,6 +17,9 @@ from dagger import Container, Platform
 
 BUILD_PLATFORMS = [Platform("linux/amd64"), Platform("linux/arm64")]
 LOCAL_BUILD_PLATFORM = Platform(f"linux/{platform.machine()}")
+
+if TYPE_CHECKING:
+    from ci_connector_ops.pipelines.bases import StepResult
 
 
 class NoBuildStepForLanguageError(Exception):
